lc0415: Add Strings
