+++ src/main/java/com/zea7ot/lc/lvl1/lc0415/followup/SolutionApproach0TwoPointers.java
@@ -1,0 +1,12 @@
+/**
+ * https://leetcode.com/problems/add-strings/
+ * 
+ * if there are decimals, how to solve it please
+ */
+package com.zea7ot.lc.lvl1.lc0415.followup;
+
+public class SolutionApproach0TwoPointers {
+    // public String addStrings(String num1, String num2) {
+
+    // }
+}
